攻速从次/每秒换算为秒/每次，与移动速度统一

+++ MakeAGame/Assets/Scripts/LubanDataLoad/LubanDataLoader.cs
@@ -164,7 +164,7 @@ namespace Game.LubanDataLoad
             so.sanCost = json.SanCost;
             so.deathFuncDescription = json.DeathDesc;
             so.cardSprite = LoadCardSprite(json);
-            so.hp = json.Hp;
+            so.hp = (int)json.Hp;
             so.attack = json.Atk;
             so.moveSpd = json.MoveSpd;
             so.defend = json.Def;
